Speed up EmbedText for similarity matrix construction (#71)
Speed up EmbedText for similarity matrix construction (#71)

- Rename SimilarityMatrix to StackedSimilarityMatrix, which is intended for use with multiple channels (eg BERT layers)
- Add new SimilarityMatrix class that accepts negative indices for OOV terms. Cosine embedding similarity is used for positive indices, while negative indices are exact matched.
- Rename existing EmbedText extractor to SlowEmbedText
- Add a new EmbedText extractor that produces output suitable for new SimilarityMatrix class. This is much more efficient, because we do not need to generate embeddings for OOV terms. Switch rerankers that use static embeddings (DRMM*, KNRM, PACRR) to use new extractor by default.

Closes #63 .

+++ capreolus/searcher/anserini.py
@@ -263,8 +263,9 @@ class StaticBM25RM3Rob04Yang19(Searcher):
         import shutil
 
         outfn = os.path.join(output_path, "static.run")
-        os.makedirs(output_path, exist_ok=True)
-        shutil.copy2(constants["PACKAGE_PATH"] / "data" / "rob04_yang19_rm3.run", outfn)
+        if not os.path.exists(outfn):
+            os.makedirs(output_path, exist_ok=True)
+            shutil.copy2(constants["PACKAGE_PATH"] / "data" / "rob04_yang19_rm3.run", outfn)
 
         return output_path
